+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Real-time token example updates in Movie Management settings when Global Management settings change (title clean rules and video dynamic range).
 - Added new settings tab **Global Management**. This tab will hold global rename settings for both Series/Movies.
+- New custom widget - ComparisonFileMatcher. This will be used to select a matching **source** file for movies/series.
 
 ### Changed